requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.0.2
django-markdownx==4.0.0b1
importlib-metadata==6.0.0
Markdown==3.4.1
Pillow==9.3.0
python-dotenv==0.21.0
sqlparse==0.4.3
svgwrite==1.4.3
Tree==0.2.4
zipp==3.11.0
